+++ include/roq/client/dispatcher.h
@@ -4,6 +4,8 @@
 
 #include "roq/api.h"
 
+#include "roq/server/trace_info.h"
+
 #include "roq/client/custom_message.h"
 
 namespace roq {
@@ -42,6 +44,35 @@ class ROQ_PUBLIC Dispatcher {
 
   //! Useful to communicate between threads
   virtual void enqueue(const CustomMessage &) = 0;
+
+  //! Make it possible to capture receive_time (origin_create_time_utc is optional)
+  template <typename Callback>
+  void create_trace_info(Callback &&callback, std::chrono::nanoseconds origin_create_time_utc = {}) {
+    TimeSetter time_setter(*this, origin_create_time_utc);
+    callback(static_cast<const server::TraceInfo &>(time_setter));
+  }
+
+ protected:
+  virtual server::TraceInfo create_trace_info(std::chrono::nanoseconds origin_create_time_utc = {}) = 0;
+  virtual void release_trace_info() = 0;
+
+  struct TimeSetter final {
+    TimeSetter(Dispatcher &dispatcher, std::chrono::nanoseconds origin_create_time_utc)
+        : dispatcher_(dispatcher), trace_info_(dispatcher_.create_trace_info(origin_create_time_utc)) {}
+
+    ~TimeSetter() {
+      try {
+        dispatcher_.release_trace_info();
+      } catch (...) {
+      }
+    }
+
+    operator const server::TraceInfo &() const { return trace_info_; }
+
+   private:
+    Dispatcher &dispatcher_;
+    server::TraceInfo trace_info_;
+  };
 };
 
 }  // namespace client