+++ include/roq/compat.h
@@ -16,10 +16,19 @@
 #define ROQ_UNLIKELY(x) (__builtin_expect(x, false))
 // c++20 uses the [[ likely ]] and [[ unlikely ]] attributes
 
+#if defined(__APPLE__)
+#if defined(__arm64__)
+#define ROQ_CACHELINE_SIZE 128
+#define ROQ_PAGE_SIZE 16384
+#else  // not __arm64__
 #define ROQ_CACHELINE_SIZE 64
+#define ROQ_PAGE_SIZE 4096
+#endif
+#else  // not __APPLE__
+#define ROQ_CACHELINE_SIZE 64
+#define ROQ_PAGE_SIZE 4096
+#endif
 // constexpr std::hardware_{constructive,destructive}_interference_size
 // not supported by gcc:
 // http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2016/p0154r1.html
 // https://gcc.gnu.org/onlinedocs/libstdc++/manual/status.html#status.iso.2017
-
-#define ROQ_PAGE_SIZE 4096