+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes will be documented in this file.
 
 ### Added
 
+* `SecurityType::SWAP`
 * Capture `origin_create_time` for externally triggered events (#140)
 
 ## 0.7.9 &ndash; 2021-12-08